% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_ms2.R
\name{combine_ms2}
\alias{combine_ms2}
\title{combine ms2 annotation object}
\usage{
combine_ms2(..., list = NULL)
}
\arguments{
\item{...}{the annotation object. Each should be a list object containing at least two elements named 'meta' and 'peakList'}

\item{list}{a list of annotation object}
}
\description{
combine ms2 annotation object
}
