% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_headers.R
\name{.getVal}
\alias{.getVal}
\title{get value from a named vector, if not, return NA}
\usage{
.getVal(x, val)
}
\arguments{
\item{x}{a named vector}

\item{val}{value to get}
}
\description{
get value from a named vector, if not, return NA
}
