% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_headers.R
\name{.msms_id_creater}
\alias{.msms_id_creater}
\title{id creater for MS2}
\usage{
.msms_id_creater(num, db)
}
\arguments{
\item{num}{the number}

\item{db}{the db prefix}
}
\description{
id creater for MS2
}
