% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_headers.R
\name{.toMonoisoMass}
\alias{.toMonoisoMass}
\title{Convert precusor MZ to monosiotopic mass}
\usage{
.toMonoisoMass(type, precMass)
}
\arguments{
\item{type}{type of adduct, such as "[M+Cl]-"}

\item{precMass}{the precursor mass}
}
\description{
Convert precusor MZ to monosiotopic mass
}
\examples{
# q <- .toMonoisoMass(type = c("[M-H]-", "[M-H]-"), precMass = c(194.91767134783, 174.97229374783))
# c(2,3,5-Trichlorphenol, 2,4-Dichloro-6-Methylphenol)

}
