% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation.R
\name{get_annotation_fragment}
\alias{get_annotation_fragment}
\title{Get annotation of fragments (MS2)}
\usage{
get_annotation_fragment(what)
}
\arguments{
\item{what}{the annotation to get, the options are:\cr 
HMDB_experimental\cr 
HMDB_predicted\cr 
MSDIAL_BioMSMS-Neg-PlaSMA\cr 
MSDIAL_BioMSMS-Pos-PlaSMA\cr 
MSDIAL_KI-GIAR_zic-HILIC_Pos_v0.90\cr 
MSDIAL_Neg-CASMI2016\cr 
MSDIAL_Neg-FiehnHILIC\cr 
MSDIAL_Neg-GNPS\cr 
MSDIAL_Neg-MassBankEU\cr 
MSDIAL_Neg-MassBank\cr 
MSDIAL_Neg-MetaboBASE\cr 
MSDIAL_Neg-PlaSMA\cr 
MSDIAL_Neg-Respect\cr 
MSDIAL_Neg-RikenOxPLs\cr 
MSDIAL_Neg-Vaniya-Fiehn_Natural_Products_Library_20200109\cr 
MSDIAL_Pos-CASMI2016\cr 
MSDIAL_Pos-FiehnHILIC\cr 
MSDIAL_Pos-GNPS\cr 
MSDIAL_Pos-MassBankEU\cr 
MSDIAL_Pos-MassBank\cr 
MSDIAL_Pos-MetaboBASE\cr 
MSDIAL_Pos-Pathogen_Box_20200109\cr 
MSDIAL_Pos-PlaSMA\cr 
MSDIAL_Pos-Respect\cr 
MSDIAL_Pos-Vaniya-Fiehn_Natural_Products_Library_20200109\cr 
MSDIAL_Public-Neg-VS14\cr 
MSDIAL_Public-Pos-VS14\cr}
}
\description{
Get annotation of fragments (MS2)
}
