% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation.R
\name{get_annotation_mass}
\alias{get_annotation_mass}
\title{Get annotation of masses}
\usage{
get_annotation_mass(what)
}
\arguments{
\item{what}{the annotation to get, the options are:
HMDB
MSDIAL_BioMSMS-Neg-PlaSMA
MSDIAL_BioMSMS-Pos-PlaSMA
MSDIAL_KI-GIAR_zic-HILIC_Pos_v0.90
MSDIAL_Neg-CASMI2016
MSDIAL_Neg-FiehnHILIC
MSDIAL_Neg-GNPS
MSDIAL_Neg-MassBankEU
MSDIAL_Neg-MassBank
MSDIAL_Neg-MetaboBASE
MSDIAL_Neg-PlaSMA
MSDIAL_Neg-Respect
MSDIAL_Neg-RikenOxPLs
MSDIAL_Neg-Vaniya-Fiehn_Natural_Products_Library_20200109
MSDIAL_Pos-CASMI2016
MSDIAL_Pos-FiehnHILIC
MSDIAL_Pos-GNPS
MSDIAL_Pos-MassBankEU
MSDIAL_Pos-MassBank
MSDIAL_Pos-MetaboBASE
MSDIAL_Pos-Pathogen_Box_20200109
MSDIAL_Pos-PlaSMA
MSDIAL_Pos-Respect
MSDIAL_Pos-Vaniya-Fiehn_Natural_Products_Library_20200109
MSDIAL_Public-Neg-VS14
MSDIAL_Public-Pos-VS14}
}
\description{
Get annotation of masses
}
