% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_annotation_hmdb.R
\name{prep_annotation_hmdb}
\alias{prep_annotation_hmdb}
\title{Prepare hmdb data for mass annotation in xcms file}
\usage{
prep_annotation_hmdb(x)
}
\arguments{
\item{x}{an xml file downloaded from hmdb database or an RDS file reading the RDS.}
}
\description{
Prepare hmdb data for mass annotation in xcms file
}
