% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_annotation_hmdb_msms.R
\name{prep_annotation_hmdb_msms}
\alias{prep_annotation_hmdb_msms}
\title{parse MSMMS XML files downloaded from HMDB}
\usage{
prep_annotation_hmdb_msms(xmlFolder, ...)
}
\arguments{
\item{xmlFolder}{the xml folder}

\item{...}{other parameters passed to bplapply}
}
\description{
parse MSMMS XML files downloaded from HMDB
}
