% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_annotation_msdial.R
\name{prep_annotation_msdial}
\alias{prep_annotation_msdial}
\title{Processing MSP files of msDial to list of meta table for ms1 and ms2 and peak list}
\usage{
prep_annotation_msdial(mspFile, db, predicted = c("false", "true")[1],
  cores = 1)
}
\arguments{
\item{mspFile}{.msp file}

\item{db}{the database prefix, used for creating IDs}

\item{predicted}{predicted spectra or expermental}

\item{cores}{passed to mclapply}
}
\description{
Processing MSP files of msDial to list of meta table for ms1 and ms2 and peak list
}
